breezed-fastapi-main/requirements.txt
fastapi[all]
fastapi_sqlalchemy
openai==1.9.0
alembic
psycopg2==2.9.3
python-dotenv
requests
twilio
sendgrid
PyJWT
stripe